+++ components/viz/service/frame_sinks/surface_synchronization_unittest.cc
@@ -2118,7 +2118,7 @@ TEST_F(SurfaceSynchronizationTest, FrameIndexWithPendingFrames) {
                           std::vector<TransferableResource>()));
   Surface* parent_surface =
       frame_sink_manager().surface_manager()->GetSurfaceForId(parent_id);
-  uint64_t initial_frame_index = parent_surface->GetActiveFrameIndex();
+  uint32_t initial_frame_index = parent_surface->GetActiveFrameIndex();
 
   // Submit frames with unresolved dependencies. GetActiveFrameIndex should
   // return the same value as before.
@@ -2175,7 +2175,7 @@ TEST_F(SurfaceSynchronizationTest, ActiveFrameIndex) {
       child_id2.local_surface_id(),
       MakeDefaultInteractiveCompositorFrame(kBeginFrameSourceId));
   EXPECT_TRUE(parent_surface()->HasActiveFrame());
-  uint64_t expected_index = kFrameIndexStart;
+  uint32_t expected_index = kFrameIndexStart;
   EXPECT_EQ(expected_index, parent_surface()->GetActiveFrameIndex());
 }
